Clinic_project_1249160/requirements.txt
Flask==2.1.0
Flask-Login==0.5.0
mysql-connector-python==8.0.28
Flask-WTF==1.0.0
Werkzeug==1.0.1
